pyproject.toml
[tool.poetry]
name = "project-vox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["all"], version = "^0.3.0"}
pydantic = {extras = ["email"], version = "^2.9.2"}
langchain-groq = "^0.2.0"
python-decouple = "^3.8"
langchain = "^0.3.2"
langchain-community = "^0.3.1"

langchain-core = "0.3.9"
groq = "^0.11.0"
psycopg2 = "^2.9.9"
matplotlib = "^3.9.2"
pandas = "^2.2.3"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.12"
bcrypt = "4.0.1"
fastapi-mail = "^1.4.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
